requirements.txt
Django==1.11.29
celery==4.2.1
django-oscar==2.0.4
django-oscar-api-checkout==0.4.1
django-bulk-update==2.2.0
python-dotenv==0.14.0
psycopg2-binary==2.8.6
requests==2.25.1

ipython==7.16.1
jedi==0.17.2